+++ src/lib/ogr.js
@@ -16,33 +16,40 @@ function createStream (format, options) {
   .on('error', e => output.emit('error', e))
   .pipe(output)
 
+  output.abort = () => ogrStream.abort()
+
   return output
 }
 
 function spawnOgr (format, options) {
-  let lastMessage
   const output = _()
   const cmd = Cmd.create(format, options)
   output.emit('log', {level: 'info', message: `Executing: OGR2OGR ${cmd.join(' ')}`})
   const ogr = spawn('ogr2ogr', cmd)
-  .on('close', c => {
-    if (c > 0) output.emit('error', new Error(`OGR Failed: ${lastMessage}`))
+  .on('close', (c, msg) => {
+    if (c > 0) output.emit('error', new Error(`OGR Failed: ${msg}`))
   })
 
   _(ogr.stderr)
+  .split()
   .each(data => {
     const msg = data.toString()
     // Error 1: GeoJSON parsing error
     // Error 4: Failed to read GeeoJSON
     // Error 6: debug message that can be ignored
-    lastMessage = msg
     output.emit('log', {level: 'error', message: msg})
     if (msg.match(/ERROR\s[^6]/)) {
       ogr.stderr.unpipe()
       ogr.kill('SIGKILL')
-      ogr.emit('close', 1)
+      ogr.emit('close', 1, msg)
     }
   })
+
+  output.abort = () => {
+    ogr.stderr.unpipe()
+    ogr.kill('SIGKILL')
+    ogr.emit('close', 1, 'Aborted')
+  }
 
   return ogr.stdout.pipe(output)
 }